pyproject.toml
[project]
name = "vtuber-plot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "openai>=1.66.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "selenium>=4.29.0",
]
